+++ src/main/java/com/app/folioman/portfolio/UserSchemeDetailService.java
@@ -1,4 +1,4 @@
-package com.app.folioman.shared;
+package com.app.folioman.portfolio;
 
 public interface UserSchemeDetailService {
